requirements.txt
Django>=4.2.0
mysql-connector-python>=8.0.0
Pillow>=9.0.0
reportlab>=3.6.0
openpyxl>=3.0.0
cryptography>=41.0.0
gunicorn>=20.1.0
whitenoise>=6.0.0
dj-database-url>=1.0.0 